requirements.txt
numpy>=1.23
pandas>=1.5
scikit-learn>=1.3
matplotlib>=3.7
joblib>=1.2
pyyaml>=6.0
